adding base test for og image url
